pyproject.toml
[build-system]
requires = ["hatchling>=1.27"]
build-backend = "hatchling.build"

[project]
name = "stampify"
version = "0.1.4"
description = "Turn black-and-white artwork into stamp-ready STL models."
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
license-files = ["LICENSE"]
keywords = ["cli", "ink", "stamp", "stl", "3d-printing"]
classifiers = [
  "Environment :: Console",
  "Intended Audience :: End Users/Desktop",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Multimedia :: Graphics :: 3D Modeling",
]
dependencies = [
  "manifold3d>=3,<4",
  "numpy>=2,<3",
  "pillow>=10,<12",
  "scikit-image>=0.24,<1",
  "shapely>=2,<3",
  "triangle>=20250106",
  "trimesh>=4,<5",
]

[project.scripts]
stampify = "ink_print.cli:main"

[project.optional-dependencies]
test = [
  "pytest>=8,<9",
]

[project.urls]
Homepage = "https://github.com/nwcell/stampify"
Repository = "https://github.com/nwcell/stampify"
Issues = "https://github.com/nwcell/stampify/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/ink_print"]

[tool.hatch.build.targets.sdist]
include = [
  "/LICENSE",
  "/README.md",
  "/image_to_stamp.py",
  "/pyproject.toml",
  "/sample",
  "/src/ink_print",
  "/tests",
]
